Fix job type error

+++ src/models/job.rs
@@ -1,4 +1,4 @@
-use sqlx::{FromRow, types::chrono};
+use sqlx::{FromRow, types::{chrono, BigDecimal}};
 
 #[derive(Debug, FromRow)]
 pub struct Job {
@@ -7,7 +7,7 @@ pub struct Job {
     pub description: Option<String>,
     pub company_id: i32,
     pub location: Option<String>,
-    pub salary: Option<f64>,
+    pub salary: Option<BigDecimal>,
     pub created_at: chrono::NaiveDateTime,
     pub expires_at: Option<chrono::NaiveDateTime>,
 }
